+++ src/routes/(app)/app/+page.svelte
@@ -7,7 +7,6 @@
 		MagicWand01Icon,
 		NoteEditIcon,
 		PanelRightCloseIcon,
-		PanelRightOpenIcon,
 		SearchList01Icon,
 		Folder01Icon,
 		Tag01Icon,
@@ -20,11 +19,19 @@
 	import { api } from '../../../convex/_generated/api.js';
 	import type { Doc, Id } from '../../../convex/_generated/dataModel';
 	import { Button } from '$lib/components/ui/button';
-	import * as Empty from '$lib/components/ui/empty';
-	import * as InputGroup from '$lib/components/ui/input-group';
+	import BookmarkList from '$lib/components/library/bookmark-list.svelte';
+	import LibraryHeader from '$lib/components/library/library-header.svelte';
 	import * as ScrollArea from '$lib/components/ui/scroll-area';
 	import { Skeleton } from '$lib/components/ui/skeleton';
 	import { Textarea } from '$lib/components/ui/textarea';
+	import {
+		formatDate,
+		getDisplayTitle,
+		getExtractionContext,
+		getHostname,
+		normalizeTagName,
+		normalizeTagNames
+	} from '$lib/bookmark-utils';
 	import { cn } from '$lib/utils';
 
 	type BookmarkRow = {
@@ -59,7 +66,6 @@
 	const rows = $derived((bookmarksResponse.data ?? []) as BookmarkRow[]);
 	const collections = $derived((collectionsResponse.data ?? []) as Doc<'collections'>[]);
 	const allTags = $derived((tagsResponse.data ?? []) as Doc<'tags'>[]);
-	const skeletonRows = [0, 1, 2, 3, 4, 5];
 	let selectedBookmarkId = $state<Id<'bookmarks'> | null>(null);
 	let url = $state('');
 	let saveSuccess = $state('');
@@ -98,61 +104,14 @@
 			: null
 	);
 
-	const statusMeta = {
-		pending: { label: 'Pending', icon: Clock01Icon },
-		enriched: { label: 'Enriched', icon: CheckmarkCircle02Icon },
-		failed: { label: 'Failed', icon: MagicWand01Icon }
+	const extractionStatusIcons = {
+		pending: Clock01Icon,
+		enriched: CheckmarkCircle02Icon,
+		failed: MagicWand01Icon
 	};
-
-	function getDisplayTitle(bookmark: Doc<'bookmarks'>) {
-		return bookmark.title?.trim() || getHostname(bookmark.url) || bookmark.url;
-	}
-
-	function getHostname(url: string) {
-		try {
-			return new URL(url).hostname.replace(/^www\./, '');
-		} catch {
-			return url;
-		}
-	}
-
-	function getExtractionContext(bookmark: Doc<'bookmarks'>) {
-		if (bookmark.extractionStatus === 'enriched') {
-			return 'Extracted text is ready for grounded answers.';
-		}
-		if (bookmark.extractionStatus === 'failed') {
-			return bookmark.extractionError || 'Extraction failed. Retry when you are ready.';
-		}
-		return 'This bookmark becomes stronger context after enrichment finishes.';
-	}
-
-	function getTags(row: BookmarkRow) {
-		return row.tags.map((tag) => tag.name);
-	}
-
-	function getCollectionName(row: BookmarkRow) {
-		return row.collection?.name ?? 'Unassigned';
-	}
-
-	function formatDate(timestamp?: number) {
-		if (!timestamp) return null;
-		return new Intl.DateTimeFormat('en', {
-			month: 'short',
-			day: 'numeric',
-			year: 'numeric'
-		}).format(timestamp);
-	}
 
 	function openSourceUrl(url: string) {
 		window.open(url, '_blank', 'noreferrer');
-	}
-
-	function normalizeTagName(name: string) {
-		return name.trim().replace(/\s+/g, ' ');
-	}
-
-	function normalizeTagNames(names: string[]) {
-		return Array.from(new Set(names.map(normalizeTagName).filter(Boolean)));
 	}
 
 	const savedTagNames = $derived(selectedTags.map((tag) => tag.name));
@@ -387,151 +346,24 @@
 	)}
 >
 	<section class="flex min-h-0 min-w-0 flex-col">
-		<div class="border-b border-border/50 px-4 py-3">
-			<div class="flex items-start justify-between gap-3">
-				<div class="min-w-0">
-					<div class="flex items-center gap-2">
-						<h1 class="text-xl leading-tight font-semibold">Library</h1>
-						<span class="text-[11px] text-muted-foreground">{rows.length} saved</span>
-					</div>
-					<p class="mt-0.5 text-xs leading-snug text-muted-foreground">
-						Capture links, triage metadata, and keep context ready for Ask.
-					</p>
-				</div>
-				<Button
-					type="button"
-					variant="ghost"
-					size="icon-sm"
-					class="hidden size-8 shrink-0 lg:inline-flex"
-					onclick={() => (isInspectorOpen = !isInspectorOpen)}
-					aria-pressed={isInspectorOpen}
-					title={isInspectorOpen ? 'Hide details' : 'Show details'}
-				>
-					<HugeiconsIcon
-						icon={isInspectorOpen ? PanelRightCloseIcon : PanelRightOpenIcon}
-						strokeWidth={2}
-					/>
-					<span class="sr-only">{isInspectorOpen ? 'Hide details' : 'Show details'}</span>
-				</Button>
-			</div>
+		<LibraryHeader
+			count={rows.length}
+			bind:url
+			{isSaving}
+			{saveError}
+			{saveSuccess}
+			{isInspectorOpen}
+			onsubmit={handleSave}
+			onToggleInspector={() => (isInspectorOpen = !isInspectorOpen)}
+		/>
 
-			<form class="mt-3" onsubmit={handleSave}>
-				<InputGroup.InputGroup class="h-8">
-					<InputGroup.InputGroupAddon>
-						<HugeiconsIcon icon={Link01Icon} strokeWidth={2} />
-					</InputGroup.InputGroupAddon>
-					<InputGroup.InputGroupInput
-						bind:value={url}
-						placeholder="Paste a URL to save..."
-						type="text"
-						inputmode="url"
-						disabled={isSaving}
-						class="text-xs"
-					/>
-					<InputGroup.InputGroupButton type="submit" size="sm" disabled={isSaving || !url.trim()}>
-						{isSaving ? 'Saving...' : 'Save'}
-					</InputGroup.InputGroupButton>
-				</InputGroup.InputGroup>
-				{#if saveError}
-					<p class="mt-2 text-xs text-destructive">{saveError}</p>
-				{:else if saveSuccess}
-					<p class="mt-2 text-xs text-muted-foreground">{saveSuccess}</p>
-				{/if}
-			</form>
-		</div>
-
-		<div
-			class="grid h-9 grid-cols-[minmax(0,1fr)_6.5rem] items-center gap-3 border-b border-border/50 px-4 text-[11px] font-medium text-muted-foreground uppercase sm:grid-cols-[minmax(0,1fr)_7rem_8rem] lg:grid-cols-[minmax(0,1fr)_7rem_8rem_12rem]"
-		>
-			<span>Bookmark</span>
-			<span>Status</span>
-			<span class="hidden sm:block">Collection</span>
-			<span class="hidden lg:block">Tags</span>
-		</div>
-
-		<ScrollArea.ScrollArea class="min-h-0 flex-1">
-			{#if bookmarksResponse.isLoading}
-				<div class="flex flex-col gap-1 p-2">
-					{#each skeletonRows as index (index)}
-						<div
-							class="grid h-11 grid-cols-[minmax(0,1fr)_6.5rem] items-center gap-3 rounded-md px-2 sm:grid-cols-[minmax(0,1fr)_7rem_8rem] lg:grid-cols-[minmax(0,1fr)_7rem_8rem_12rem]"
-						>
-							<div class="space-y-2">
-								<Skeleton class="h-3.5 w-2/3" />
-								<Skeleton class="h-3 w-1/3" />
-							</div>
-							<Skeleton class="h-4 w-16" />
-							<Skeleton class="hidden h-4 w-20 sm:block" />
-							<Skeleton class="hidden h-5 w-28 lg:block" />
-						</div>
-					{/each}
-				</div>
-			{:else if bookmarksResponse.error}
-				<div class="p-4">
-					<p class="text-sm text-destructive">
-						Bookmarks could not load. {bookmarksResponse.error.message}
-					</p>
-				</div>
-			{:else if rows.length === 0}
-				<div class="flex min-h-96 p-4">
-					<Empty.Empty class="border-0">
-						<Empty.EmptyHeader>
-							<Empty.EmptyMedia>
-								<HugeiconsIcon icon={Link01Icon} strokeWidth={2} />
-							</Empty.EmptyMedia>
-							<Empty.EmptyTitle>No bookmarks yet</Empty.EmptyTitle>
-							<Empty.EmptyDescription>
-								Paste a URL above to start building your library.
-							</Empty.EmptyDescription>
-						</Empty.EmptyHeader>
-					</Empty.Empty>
-				</div>
-			{:else}
-				<div class="flex flex-col gap-1 p-2">
-					{#each rows as row (row.bookmark._id)}
-						{@const status = statusMeta[row.bookmark.extractionStatus]}
-						{@const tags = getTags(row)}
-						<button
-							type="button"
-							onclick={() => (selectedBookmarkId = row.bookmark._id)}
-							class={cn(
-								'grid h-11 grid-cols-[minmax(0,1fr)_6.5rem] items-center gap-3 rounded-md px-2 text-left transition-colors hover:bg-accent sm:grid-cols-[minmax(0,1fr)_7rem_8rem] lg:grid-cols-[minmax(0,1fr)_7rem_8rem_12rem]',
-								selectedRow?.bookmark._id === row.bookmark._id && 'bg-accent text-accent-foreground'
-							)}
-						>
-							<span class="min-w-0">
-								<span class="block truncate text-xs leading-tight font-medium">
-									{getDisplayTitle(row.bookmark)}
-								</span>
-								<span class="mt-0.5 block truncate text-[11px] leading-tight text-muted-foreground">
-									{row.bookmark.siteName || getHostname(row.bookmark.url)}
-								</span>
-							</span>
-							<span class="flex min-w-0 items-center gap-1.5 text-[11px] text-muted-foreground">
-								<HugeiconsIcon icon={status.icon} strokeWidth={2} class="size-3.5 shrink-0" />
-								<span class="truncate">{status.label}</span>
-							</span>
-							<span class="hidden truncate text-[11px] text-muted-foreground sm:block">
-								{getCollectionName(row)}
-							</span>
-							<span class="hidden min-w-0 items-center gap-1.5 lg:flex">
-								{#if tags.length}
-									{#each tags.slice(0, 3) as tag (tag)}
-										<span
-											class="max-w-24 truncate rounded-sm bg-secondary px-1.5 py-0.5 text-[11px] text-secondary-foreground"
-										>
-											{tag}
-										</span>
-									{/each}
-								{:else}
-									<span class="text-[11px] text-muted-foreground">No tags</span>
-								{/if}
-							</span>
-						</button>
-					{/each}
-				</div>
-			{/if}
-		</ScrollArea.ScrollArea>
+		<BookmarkList
+			{rows}
+			isLoading={bookmarksResponse.isLoading}
+			error={bookmarksResponse.error}
+			selectedBookmarkId={selectedRow?.bookmark._id ?? null}
+			onselect={(bookmarkId) => (selectedBookmarkId = bookmarkId)}
+		/>
 	</section>
 
 	<aside
@@ -741,7 +573,7 @@
 					<section>
 						<div class="flex items-center gap-2">
 							<HugeiconsIcon
-								icon={statusMeta[selectedBookmark.extractionStatus].icon}
+								icon={extractionStatusIcons[selectedBookmark.extractionStatus]}
 								strokeWidth={2}
 								class="size-3.5 text-muted-foreground"
 							/>
